feature: video block (#344)
* feat: video block

* fix: switch condition

* fix: pr change

+++ apps/journeys-admin/src/components/Editor/VideoBlockEditor/VideoBlockEditor.tsx
@@ -16,14 +16,16 @@ interface VideoBlockEditorProps {
   selectedBlock: TreeBlock<VideoBlock> | null
   parentBlockId: string
   parentOrder: number
+  showDelete?: boolean
   onChange: (block: TreeBlock<VideoBlock>) => Promise<void>
-  onDelete: () => Promise<void>
+  onDelete?: () => Promise<void> | undefined
 }
 
 export function VideoBlockEditor({
   selectedBlock,
   parentBlockId,
   parentOrder,
+  showDelete = true,
   onChange,
   onDelete
 }: VideoBlockEditorProps): ReactElement {
@@ -40,7 +42,7 @@ export function VideoBlockEditor({
 
   const handleVideoDelete = async (): Promise<void> => {
     setTabValue(0)
-    await onDelete()
+    if (onDelete != null) await onDelete()
   }
 
   return (
@@ -55,7 +57,7 @@ export function VideoBlockEditor({
         caption={
           selectedBlock?.videoContent?.src == null ? 'Formats: MP4, HLS' : ''
         }
-        showDelete={selectedBlock?.videoContent?.src != null}
+        showDelete={showDelete && selectedBlock?.videoContent?.src != null}
         onDelete={handleVideoDelete}
       />
       <Box
